fix: move DummyStorageTransformer to zarr, and import it from tests instead of the other way around
